Tim and I fixed the timer.
We had to add a super class called question and created a new method in each of the questions to get the state, also at the behest of Kevin we added a method restart to the Timer class which saves 2 whole lines per question class. We also messed with the question stack and mouselistener to make it work correctly.